Implementação da DTO, correção dos tipos primitivos Long, Implementação da Estabelecimentorepository

+++ src/main/java/com/controle_de_gastos/notas_api/model/ItemTipo.java
@@ -11,6 +11,6 @@ import lombok.*;
 public class ItemTipo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long idTipo;
+    private Integer idTipo;
     private String nome;
 }
